\name{sim.QTL.data.GxE}
\alias{sim.QTL.data.GxE}
\title{Simulates subjects for continuous outcome}
\usage{
sim.QTL.data.GxE(n = NULL, ph.mean = NULL, ph.sd = NULL, e1.efkt = NULL,
  e1.prev = NULL, e1.mean = NULL, e1.sd = NULL, e1.low.lim = NULL,
  e1.up.lim = NULL, e2.efkt = NULL, e2.prev = NULL, e2.mean = NULL,
  e2.sd = NULL, e2.low.lim = NULL, e2.up.lim = NULL, i.efkt = NULL,
  pheno.rel = NULL)
}
\arguments{
  \item{n}{Number of subjects to simulate}

  \item{ph.mean}{statistical mean}

  \item{e1.model}{Model of the 1st environmental exposure}

  \item{e1.efkt}{Effects of the 1st environmental exposure}

  \item{e1.prev}{Prevalence of the 1st environmental
  exposure}

  \item{e1.mean}{Mean of the 1st environmental exposure}

  \item{e1.sd}{Standard deviation of the 1st environmental
  exposure}

  \item{e1.low.lim}{lower limit of the 1st environmental
  exposure}

  \item{e1.up.lim}{upper limit of the 1st environmental
  exposure}

  \item{e2.model}{Model of the 2nd environmental exposure}

  \item{e2.efkt}{Effects of the 2nd environmental exposure}

  \item{e2.prev}{Prevalence of the 2nd environmental
  exposure}

  \item{e2.mean}{Mean of the 2nd environmental exposure}

  \item{e2.sd}{Standard deviation of the 2nd environmental
  exposure}

  \item{e2.low.lim}{lower limit of the 2nd environmental
  exposure}

  \item{e2.up.lim}{upper limit of the 2nd environmental
  exposure}

  \item{i.efkt}{Interaction effect}

  \item{pheno.rel}{reliability of the assessment for a
  quantitative outcome.}
}
\value{
A matrix
}
\description{
Generates the specified number of subjects
}
\details{

}
\author{
Gaye A.
}
\keyword{internal}

